integrated modelmapper for easy conversion between entitiy's and dto's

+++ src/main/java/nl/novi/les/springboot/backendtechiteasy/services/TelevisionService.java
@@ -5,9 +5,11 @@ import nl.novi.les.springboot.backendtechiteasy.models.dtos.TelevisionCreatedDto
 import nl.novi.les.springboot.backendtechiteasy.exceptions.RecordNotFoundException;
 import nl.novi.les.springboot.backendtechiteasy.models.entities.Television;
 import nl.novi.les.springboot.backendtechiteasy.repositories.TelevisionRepository;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,14 +17,17 @@ import java.util.Optional;
 public class TelevisionService {
 
     private final TelevisionRepository televisionRepository;
+    private final ModelMapper modelMapper = new ModelMapper();
 
     @Autowired
     public TelevisionService(TelevisionRepository televisionRepository) {
         this.televisionRepository = televisionRepository;
     }
 
-    public List<Television> getAllTvs() {
-        return televisionRepository.findAll();
+    public List<TelevisionCreatedDto> getAllTvs() {
+        List<Television> televisionList = televisionRepository.findAll();
+
+        return convertTelevisionListToDtoList(televisionList);
     }
 
     public TelevisionCreatedDto getTvById(Long id) {
@@ -32,39 +37,20 @@ public class TelevisionService {
         }
 
         Television tv = televisionRepository.getById(id);
-        TelevisionCreatedDto tvDto = new TelevisionCreatedDto();
-        tvDto.setId(tv.getId());
-        tvDto.setType(tv.getType());
-        tvDto.setBrand(tv.getBrand());
-        tvDto.setName(tv.getName());
-        tvDto.setPrice(tv.getPrice());
 
-        return tvDto;
+        return modelMapper.map(tv, TelevisionCreatedDto.class);
     }
 
     public TelevisionCreatedDto addTv(TelevisionDto newTv) {
 
-        Television tv = new Television();
-
-        tv.setType(newTv.getType());
-        tv.setBrand(newTv.getBrand());
-        tv.setName(newTv.getName());
-        tv.setPrice(newTv.getPrice());
+        Television tv = modelMapper.map(newTv, Television.class);
 
         tv = televisionRepository.save(tv);
 
-        TelevisionCreatedDto createdTelevision = new TelevisionCreatedDto();
-
-        createdTelevision.setId(tv.getId());
-        createdTelevision.setType(tv.getType());
-        createdTelevision.setBrand(tv.getBrand());
-        createdTelevision.setName(tv.getName());
-        createdTelevision.setPrice(tv.getPrice());
-
-        return createdTelevision;
+        return modelMapper.map(tv, TelevisionCreatedDto.class);
     }
 
-    public Television updateTv(long id, Television tv) {
+    public TelevisionCreatedDto updateTv(long id, TelevisionCreatedDto tv) {
 
         Optional<Television> result = televisionRepository.findById(id);
 
@@ -72,27 +58,53 @@ public class TelevisionService {
             throw new RecordNotFoundException("tv cannot be found!");
         }
 
-        return televisionRepository.save(tv);
+        Television updatedTv = modelMapper.map(tv, Television.class);
+
+        updatedTv = televisionRepository.save(updatedTv);
+
+        return modelMapper.map(updatedTv, TelevisionCreatedDto.class);
     }
 
     public void removeTv(long id) {
+        Optional<Television> result = televisionRepository.findById(id);
+        if(result.isEmpty()) {
+            throw new RecordNotFoundException("Tv not found!");
+        }
         televisionRepository.deleteById(id);
     }
 
-    public List<Television> getAllTvsByType(String value) {
-        return televisionRepository.findTelevisionsByTypeEqualsIgnoreCase(value);
+    public List<TelevisionCreatedDto> getAllTvsByType(String value) {
+        List<Television> televisionList = televisionRepository.findTelevisionsByTypeEqualsIgnoreCase(value);
+
+        return convertTelevisionListToDtoList(televisionList);
     }
 
-    public List<Television> getAllTvsByBrand(String value) {
-        return televisionRepository.findTelevisionsByBrandEqualsIgnoreCase(value);
+    public List<TelevisionCreatedDto> getAllTvsByBrand(String value) {
+        List<Television> televisionList = televisionRepository.findTelevisionsByBrandEqualsIgnoreCase(value);
+
+        return convertTelevisionListToDtoList(televisionList);
     }
 
-    public List<Television> getAllTvsByName(String value) {
-        return televisionRepository.findTelevisionsByNameEqualsIgnoreCase(value);
+    public List<TelevisionCreatedDto> getAllTvsByName(String value) {
+        List<Television> televisionList = televisionRepository.findTelevisionsByNameEqualsIgnoreCase(value);
+
+        return convertTelevisionListToDtoList(televisionList);
     }
 
-    public List<Television> getAllTvsByPrice(double value) {
-        return televisionRepository.findTelevisionsByPriceIs(value);
+    public List<TelevisionCreatedDto> getAllTvsByPrice(double value) {
+        List<Television> televisionList = televisionRepository.findTelevisionsByPriceIs(value);
+
+        return convertTelevisionListToDtoList(televisionList);
+    }
+
+    private List<TelevisionCreatedDto> convertTelevisionListToDtoList(List<Television> televisionList) {
+        List<TelevisionCreatedDto> returnList = new ArrayList<>();
+
+        for(Television tv : televisionList) {
+            returnList.add(modelMapper.map(tv, TelevisionCreatedDto.class));
+        }
+
+        return returnList;
     }
 
 }
